requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-dotenv==0.19.2
six==1.16.0
sqlparse==0.4.2
whitenoise==6.0.0
